improve cross DB queries
