+++ twitter_patches/null.rb
@@ -1,0 +1,13 @@
+module Tws
+  module TwitterPatches
+    module Null
+      def present?
+        false
+      end
+ 
+      def nil?
+        true
+      end
+    end
+  end
+end
